Move shouldBeProofValue to assertions.
